Add per-row and end constructs, rename init block (#5)
Rename init block to use `# pysvglabel: init` instead of emojis. Snakes
stay there because of the need to be compact.

Adds a per row code that runs before label processing (using `#
pysvglabel: row`) and a end-of-file code that runs after the last label
(using `# pysvglabel: end`). End code can be used to validate that all
data of some type was processed, in conjunction with the per-row code.